projects/wpp_ai_0001/ai-api/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.9
redis==5.0.1
django-redis==5.4.0
Pillow==10.1.0
python-magic==0.4.27
openai==1.3.0
requests==2.31.0
python-decouple==3.8
django-cors-headers==4.3.1